+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# dev
+- replace `pdal info --metadata` subprocess by a python command (used in the `color` module in particular)
+
 # 1.4.1
 - fix copy_and_hack_decorator (was not returning the decorated function output)